fix(package): move dependecies to peerDependencies in package.json and update explorer to use discoveryService instead of moduleConainer
