+++ package.json
@@ -28,9 +28,9 @@
     "url": "https://github.com/behrad-kzm/kafka-nestjs/issues"
   },
   "homepage": "https://github.com/behrad-kzm/kafka-nestjs#readme",
-  "dependencies": {
-    "@nestjs/common": "^11.1.2",
-    "@nestjs/core": "^11.1.2",
+  "peerDependencies": {
+    "@nestjs/common": "^10.0.0",
+    "@nestjs/core": "^10.0.0",
     "kafkajs": "^2.2.4"
   },
   "devDependencies": {
