+++ src/index.ts
@@ -1,20 +1,20 @@
-import { KafkaModule } from './kafka.module';
+
 import { KafkaProducerService } from './services/kafka-producer.service';
 import { KafkaExplorer } from './services/kafka.explorer';
 import { SharedKafkaAsyncConfiguration } from './interfaces/shared-queue-async-configuration';
 import { KafkaDuplicateGroupIdServerException } from './exceptions/kafka.duplicate-group-id.server.exception';
-import { KafkaConsumerOptions, KafkaConsumer } from './decorators/kafka.decorator';
+import { KafkaConsumerOptions, KafkaConsumer, KafkaProcessor } from './decorators/kafka.decorator';
 import { KafkaInvalidConfigServerException } from './exceptions/kafka.invalid-config.server.exception';
-import { KafkaProcessor } from './decorators/kafka-processor.decorator';
+import { KafkaModule } from './kafka.module';
 
 export {
-  KafkaModule,
   KafkaProducerService,
-  KafkaExplorer as KafkaDynamicListenerService,
+  KafkaExplorer,
   SharedKafkaAsyncConfiguration,
   KafkaInvalidConfigServerException,
   KafkaDuplicateGroupIdServerException,
   KafkaConsumerOptions,
   KafkaConsumer,
   KafkaProcessor,
+  KafkaModule
 };
