+++ src/interfaces/shared-queue-async-configuration.ts
@@ -1,7 +1,7 @@
-import { FactoryProvider } from '@nestjs/common';
+import { FactoryProvider, ModuleMetadata } from '@nestjs/common';
 import { KafkaConfig } from 'kafkajs';
 
-export interface SharedKafkaAsyncConfiguration {
+export interface SharedKafkaAsyncConfiguration extends Pick<ModuleMetadata, 'imports'> {
   useFactory?: (
     ...args: any[]
   ) => Promise<KafkaConfig> | KafkaConfig;
